[refactor]: 일괄저장 및 저장 오류 수정
